Fix subscribe before publish with AMQP
Subscribing is a process of creating a queue, then binding it to an
exchange. Publishing doesn't actually involve queues, AMQP publishes to
an exchange. Since the exchange was created in createPubQueue(), and
attempt to subscribe before publish was erroring on exchange not found.
Now both subscribe and publish create the exchange.