Prototype use of STOMP wildcard support
It doesn't work on the subscribe side, because when the messages come
in, the queue name doesn't match any queue known to node-stomp-client.
Basically, the match algorithm needs to be implemented in
node-stomp-client, which means the publish side topic should be, too,
and this is too big a change for now.